[necogcp] Refactor
