+++ gcp/public/etc/apt/apt.conf.d/20auto-upgrades
@@ -1,0 +1,2 @@
+APT::Periodic::Update-Package-Lists "0";
+APT::Periodic::Unattended-Upgrade "0";
