+++ gcp/public/etc/containers/registries.conf
@@ -1,0 +1,8 @@
+[registries.search]
+registries = ['docker.io']
+
+[registries.insecure]
+registries = []
+
+[registries.block]
+registries = []
